Lab 22
This is an ultra awesome function that compiles ARI value (reading level) for a given text document, in this case, we use the Gettysburg Address, it asks for the filename at the start, for no reason other than to print it at the end.  Function uses some interesting criteria, but it works!